Update discoverSingle to check for executable plugins
The is plugin executable check needs to happen before the plugin match gets added to the discovered plugins map.
In a situation where a valid plugin (executable) and a non-executable file named like the plugin are found in the
same path the non-executable file path will get added to the map and override any previous matches. By moving
the executable check into the discoverSingle method we ensure that only valid executable plugins get added a potential match.