Update test cases
On Windows the exec.LookPath just checks that a file ends with the appropriate extension
so this fails the test because the mock file is executed when it shouldn't be. In the realworld,
this is an error but for files that don't have a proper extension Packer should ignore them.

For Unix there is no extension so it is valid to check if the file is executable.